requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
transformers
torchaccelerate==1.0.1
